Initial commit.
Requires a lot of work but the basics work. Will start a node server.
Only param acceted is in the form of a url without the prefix.
install requirements, then proceed to the http server dished out.
example : http://localhost/google.com
This will infact display a screenshot of google.com for you.
This is useable in css background-image() and html image src="$n"
Enjoy